makes some small cleanup changes
