+++ PhotoPoints/Controller/StartupManager.swift
