+++ PhotoPoints/Helpers/ViewAnchor.swift
@@ -11,6 +11,19 @@ import UIKit
 
 extension UIView {
     
+    func pin(to view: UIView, padding: CGFloat = 0) {
+        self.anchor(
+            top: view.topAnchor,
+            left: view.leftAnchor,
+            bottom: view.bottomAnchor,
+            right: view.rightAnchor,
+            paddingTop: padding,
+            paddingLeft: padding,
+            paddingBottom: padding,
+            paddingRight: padding
+        )
+    }
+    
     func anchor(top: NSLayoutYAxisAnchor? = nil, left: NSLayoutXAxisAnchor? = nil, bottom: NSLayoutYAxisAnchor? = nil, right: NSLayoutXAxisAnchor? = nil, centerX: NSLayoutXAxisAnchor? = nil, centerY: NSLayoutYAxisAnchor? = nil, paddingTop: CGFloat? = 0, paddingLeft: CGFloat? = 0, paddingBottom: CGFloat? = 0, paddingRight: CGFloat? = 0, width: CGFloat? = nil, height: CGFloat? = nil) {
         
         translatesAutoresizingMaskIntoConstraints = false
