+++ PhotoPoints/View/Points/Table/DateView.swift
@@ -40,13 +40,3 @@ class DateView: UIView {
     }
     
 }
-
-extension Date {
-    func headerStyle() -> String {
-        let formatter = DateFormatter()
-        formatter.timeStyle = .none
-        formatter.dateStyle = .full
-        let dayString = String(formatter.string(from: Date()).uppercased().dropLast(6))
-        return dayString
-    }
-}
